requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
pytest==5.4.1
pytest-assume==2.2.1